refactor: weight をグラム単位、speed を 0.1km/h 単位の整数型に変更
- weight: 小数点を防ぐためグラム単位で保存（フロント側でkgに変換）
- speed: Decimal型からInt型に変更し、0.1km/h単位で保存
  - 例: 10.5 km/h → 105
  - レスポンスが文字列になる問題を解消
- サービス層からDecimal変換を削除
- ルート層からNumber変換を削除
- バリデーター、ドキュメント、スキーマのコメントを更新

+++ src/validators/training-session.ts
@@ -13,12 +13,12 @@ import { exerciseSchema } from "./exercise";
 export const workoutSetSchema = z.object({
   id: z.string(),
   // 筋トレ用
-  weight: z.number().int().min(0).nullable(),
+  weight: z.number().int().min(0).nullable(), // 重量（グラム）- フロント側で kg に変換
   reps: z.number().int().min(0).nullable(),
   // 有酸素用
   distance: z.number().int().min(0).nullable(), // 距離（メートル）
   duration: z.number().int().min(0).nullable(), // 時間（秒）
-  speed: z.number().min(0).nullable(), // 速さ（km/h）
+  speed: z.number().int().min(0).nullable(), // 速さ（0.1 km/h 単位）- 例: 10.5 km/h → 105
   calories: z.number().int().min(0).nullable(), // カロリー（kcal）
 });
 
@@ -31,13 +31,13 @@ export const workoutSetSchema = z.object({
 export const workoutSetRequestSchema = z.object({
   id: z.string().optional(),
   // 筋トレ用
-  weight: z.number().int().min(0).nullable().optional(),
+  weight: z.number().int().min(0).nullable().optional(), // 重量（グラム）
   reps: z.number().int().min(0).nullable().optional(),
   // 有酸素用
-  distance: z.number().int().min(0).nullable().optional(),
-  duration: z.number().int().min(0).nullable().optional(),
-  speed: z.number().min(0).nullable().optional(),
-  calories: z.number().int().min(0).nullable().optional(),
+  distance: z.number().int().min(0).nullable().optional(), // 距離（メートル）
+  duration: z.number().int().min(0).nullable().optional(), // 時間（秒）
+  speed: z.number().int().min(0).nullable().optional(), // 速さ（0.1 km/h 単位）
+  calories: z.number().int().min(0).nullable().optional(), // カロリー（kcal）
 });
 
 /**
